Don't retry command, that was killed
Previously, if command was sent a signal and terminated on its own,
exec-retry would treat this as a failed command and execute it again.
This is not an intended behavior, so that commit fixes it.

If err has killed = true and a signal equals 'SIGTERM', it was killed
on purpose.